Make instanceDir and userid as dispatcher creation options
