+++ ts/packages/cli/src/commands/data/add.ts
@@ -13,8 +13,8 @@ import {
     getCacheFactory,
     createActionConfigProvider,
     getSchemaNamesForActionConfigProvider,
-    getInstanceDir,
 } from "agent-dispatcher/internal";
+import { getInstanceDir } from "agent-dispatcher/helpers/data";
 import chalk from "chalk";
 import { getDefaultExplainerName } from "agent-cache";
 import { getChatModelMaxConcurrency, getChatModelNames } from "aiclient";
