+++ ts/packages/cli/src/commands/run/request.ts
@@ -6,16 +6,17 @@ import { createDispatcher } from "agent-dispatcher";
 import {
     createActionConfigProvider,
     getCacheFactory,
-    getInstanceDir,
     getSchemaNamesForActionConfigProvider,
 } from "agent-dispatcher/internal";
+import { getClientId, getInstanceDir } from "agent-dispatcher/helpers/data";
 import { getDefaultAppAgentProviders } from "default-agent-provider";
 import chalk from "chalk";
 import { getChatModelNames } from "aiclient";
 import { readFileSync, existsSync } from "fs";
 
 const modelNames = await getChatModelNames();
-const defaultAppAgentProviders = getDefaultAppAgentProviders(getInstanceDir());
+const instanceDir = getInstanceDir();
+const defaultAppAgentProviders = getDefaultAppAgentProviders(instanceDir);
 const schemaNames = getSchemaNamesForActionConfigProvider(
     await createActionConfigProvider(defaultAppAgentProviders),
 );
@@ -69,8 +70,9 @@ export default class RequestCommand extends Command {
                 ? { enabled: true, name: flags.explainer }
                 : { enabled: false },
             cache: { enabled: false },
-            persist: true,
+            persistDir: instanceDir,
             dblogging: true,
+            clientId: getClientId(),
         });
         await dispatcher.processCommand(
             `@dispatcher request ${args.request}`,
